debugger work in progress - thread/threadgroup/string instances

+++ plugins/debugger/src/main/java/org/robovm/debugger/utils/bytebuffer/ByteBufferReader.java
@@ -104,9 +104,9 @@ public abstract class ByteBufferReader {
         return res;
     }
 
-    public int readInt16() {
+    public short readInt16() {
         expects(2);
-        return (int) byteBuffer.getShort();
+        return byteBuffer.getShort();
     }
 
     public int readUnsignedInt16() {
